% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basictabler.R
\name{basictabler}
\alias{basictabler}
\title{Render a table as a HTML widget.}
\usage{
basictabler(bt, width = NULL, height = NULL, styleNamePrefix = NULL)
}
\arguments{
\item{bt}{The table to render.}

\item{width}{The target width.}

\item{height}{The target height.}

\item{styleNamePrefix}{A text prefix to be prepennded to the CSS declarations
(to ensure uniqueness).}
}
\value{
A HTML widget.
}
\description{
The \code{basictabler} function is primarily intended for use with Shiny web
applications.
}
\examples{
# See the Shiny vignette in this package for examples.
}
