% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basictabler.R
\name{basictablerOutput}
\alias{basictablerOutput}
\title{Standard function for Shiny scaffolding.}
\usage{
basictablerOutput(outputId, width = "100\%", height = "100\%")
}
\arguments{
\item{outputId}{The id of the html element that will contain the htmlwidget.}

\item{width}{The target width of the htmlwidget.}

\item{height}{The target height of the htmlwidget.}
}
\description{
Standard function for Shiny scaffolding.
}
\examples{
# See the Shiny vignette in this package for an example.
}
