% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{basictablerSample}
\alias{basictablerSample}
\title{Generate a sample basic table.}
\usage{
basictablerSample(value)
}
\value{
Example basic table describing the performance of trains run by
  different train operating companies in Birmingham (UK).
}
\description{
The \code{basictablerSample} function generates a sample basic table.
}
\examples{
# Generate sample table.
basictablerSample()
}
