% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableThemes.R
\name{getBlankTblTheme}
\alias{getBlankTblTheme}
\title{Get an empty theme for applying no styling to a table.}
\usage{
getBlankTblTheme(parentTable, themeName = "blank")
}
\arguments{
\item{parentTable}{Owning table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A TableStyles object.
}
\description{
Get an empty theme for applying no styling to a table.
}
