% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableThemes.R
\name{getCompactTheme}
\alias{getCompactTheme}
\title{Get the compact theme for styling a table.}
\usage{
getCompactTheme(parentTable, themeName = "compact")
}
\arguments{
\item{parentTable}{Owning table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A TableStyles object.
}
\description{
Get the compact theme for styling a table.
}
