% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableThemes.R
\name{getDefaultTheme}
\alias{getDefaultTheme}
\title{Get the default theme for styling a table.}
\usage{
getDefaultTheme(parentTable, themeName = "default")
}
\arguments{
\item{parentTable}{Owning table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A TableStyles object.
}
\description{
Get the default theme for styling a table.
}
