% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{getFtBorderFromCssBorder}
\alias{getFtBorderFromCssBorder}
\title{Convert CSS border values to those used by the flextable package.}
\usage{
getFtBorderFromCssBorder(text)
}
\arguments{
\item{text}{The border declaration to parse.}
}
\value{
A list containing two elements: width, style and color.
}
\description{
\code{getXlBorderFromCssBorder} parses the CSS combined border declarations
(i.e. border, border-left, border-right, border-top, border-bottom) and
returns a list containing an openxlsx border style and color as separate
elements.
}
