% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{getFtBorderWidthFromCssBorder}
\alias{getFtBorderWidthFromCssBorder}
\title{Convert CSS border width to those used by the flextable package.}
\usage{
getFtBorderWidthFromCssBorder(border)
}
\arguments{
\item{border}{A list containing an element named style.}
}
\value{
A flextable border style.
}
\description{
\code{getFtBorderStyleFromCssBorder} takes border parameters expressed as a
list (must contain an element named style) and returns a border style that
is compatible with the flextable package.
}
