% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableThemes.R
\name{getLargePlainTheme}
\alias{getLargePlainTheme}
\title{Get the large plain theme for styling a table.}
\usage{
getLargePlainTheme(parentTable, themeName = "largeplain")
}
\arguments{
\item{parentTable}{Owning table.}

\item{themeName}{The name to use as the new theme name.}
}
\value{
A TableStyles object.
}
\description{
Get the large plain theme for styling a table.
}
