% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{parseFtBorder}
\alias{parseFtBorder}
\title{Parse an ft-border value.}
\usage{
parseFtBorder(text)
}
\arguments{
\item{text}{The border declaration to parse.}
}
\value{
A list containing two elements: width, style and color.
}
\description{
\code{parseFtBorder} parses the combined ft border declarations (i.e.
ft-border, ft-border-left, ft-border-right, ft-border-top, ft-border-bottom)
and returns a list containing width, style and color as separate elements.
}
