% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styleParsing.R
\name{PxToPt}
\alias{PxToPt}
\title{Convert a number of pixels to points}
\usage{
PxToPt(px)
}
\arguments{
\item{px}{The number of pixels convert.}
}
\value{
The corresponding number of points.
}
\description{
\code{PxToPt} converts pixels to points.
1inch = 72pt = 96px according to W3C.
Ref:  https://www.w3.org/TR/css3-values/#absolute-lengths
}
