% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basictabler.R
\name{renderBasictabler}
\alias{renderBasictabler}
\title{Standard function for Shiny scaffolding.}
\usage{
renderBasictabler(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{The R expression to execute and render in the Shiny web application.}

\item{env}{Standard shiny argument for a render function.}

\item{quoted}{Standard shiny argument for a render function.}
}
\description{
Standard function for Shiny scaffolding.
}
