% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableCells.R
\docType{class}
\name{TableCells}
\alias{TableCells}
\title{A class that contains the cells from a table.}
\format{\code{\link{R6Class}} object.}
\usage{
TableCells
}
\value{
Object of \code{\link{R6Class}} with properties and methods relating
  to the cells of a table.
}
\description{
The TableCells class contains all of the TableCell objects that comprise a
table.
}
\section{Fields}{

\describe{
\item{\code{parentTable}}{Owning table.}

\item{\code{rows}}{The rows of cells in the table.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For more complete explanations and examples please see
  the extensive vignettes supplied with this package.}
  \item{\code{new(...)}}{Create a new set of table cells, specifying
  the field values documented above.}

  \item{\code{reset())}}{Clears and removes all of the cells.}
  \item{\code{getCell(r, c))}}{Get the TableCell at the specified row and
  column coordinates in the table.}
  \item{\code{addCell(r, c, cellType="cell", rawValue=NULL,
  formattedValue=NULL)}}{Add a new cell to the table.}
  \item{\code{addBlankCell(r, c, cellType="cell", visible=TRUE)}}{Add an
  empty cell to the table.}
  \item{\code{deleteCell(r, c)}}{Remove a cell from the table (replacing it
  with a blank one).}
  \item{\code{setValue(r, c, rawValue=NULL, formattedValue=NULL)}}{Set the
  value of a cell.}
  \item{\code{extendCells(rowCount=NULL, columnCount=NULL)}}{.}
  \item{\code{setCell(r, c, cell))}}{Set the TableCell at the specified row
  and column coordinates in the table.}
  \item{\code{insertRow(rowNumber=NULL)}}{Insert a new row (moving the rows
  underneath down).}
  \item{\code{deleteRow(rowNumber=NULL)}}{Delete a row (moving the rows
  underneath up.}
  \item{\code{insertColumn(columnNumber=NULL)}}{Insert a new column (moving
  other columns rightwards.}
  \item{\code{deleteColumn(columnNumber=NULL)}}{Delete a column (moving other
  columns leftwards.}
  \item{\code{getCells(specifyCellsAsList=FALSE, rowNumbers=NULL,
  columnNumbers=NULL, cellCoordinates=NULL)}}{Retrieve cells by a combination
  of row and/or column numbers.}
  \item{\code{findCells(rowNumbers=NULL, columnNumbers=NULL, minValue=NULL,
  maxValue=NULL, exactValues=NULL, includeNull=TRUE, includeNA=TRUE)}}{Find
  cells matching the specified criteria.}
  \item{\code{getColumnWidths())}}{Retrieve the width of the longest value
  (in characters) in each column.}
  \item{\code{asList())}}{Get a list representation of the table
  cells.}
  \item{\code{asJSON()}}{Get a JSON representation of the table cells.}
}
}

\examples{
# This class should only be created by the table.
# It is not intended to be created outside of the table.
}
\keyword{datasets}
