% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableHtmlRenderer.R
\docType{class}
\name{TableHtmlRenderer}
\alias{TableHtmlRenderer}
\title{A class that renders a table in HTML.}
\format{\code{\link{R6Class}} object.}
\usage{
TableHtmlRenderer
}
\value{
Object of \code{\link{R6Class}} with properties and methods that
  render to HTML.
}
\description{
The TableHtmlRenderer class creates a HTML representation of a table.
}
\section{Fields}{

\describe{
\item{\code{parentTable}}{Owning table.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For more complete explanations and examples please see
  the extensive vignettes supplied with this package.}
  \item{\code{new(...)}}{Create a new table renderer, specifying the
  field value documented above.}

  \item{\code{getTableHtml(styleNamePrefix=NULL)}}{Get a HTML representation
  of the table.}
}
}

\examples{
# This class should not be used by end users.  It is an internal class
# created only by the BasicTable class.  It is used when rendering to HTML.
# See the package vignettes for more information about outputs.
library(basictabler)
tbl <- qtbl(data.frame(a=1:2, b=3:4))
tbl$renderTable()
}
\keyword{datasets}
