% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableThemes.R
\name{getTblTheme}
\alias{getTblTheme}
\title{Get a built-in theme for styling a table.}
\usage{
getTblTheme(parentTable, themeName = NULL)
}
\arguments{
\item{parentTable}{Owning table.}

\item{themeName}{The name of the theme to retrieve.}
}
\value{
A TableStyles object.
}
\description{
\code{getTblTheme} returns the specified theme.
}
