% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableCells.R
\docType{class}
\name{TableCells}
\alias{TableCells}
\title{A class that contains the cells from a table.}
\format{\code{\link{R6Class}} object.}
\usage{
TableCells
}
\value{
Object of \code{\link{R6Class}} with properties and methods relating
  to the cells of a table.
}
\description{
The TableCells class contains all of the TableCell objects that comprise a
table.
}
\section{Fields}{

\describe{
\item{\code{parentTable}}{Owning pivot table.}

\item{\code{rows}}{The rows of cells in the table.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For more complete explanations and examples please see
  the extensive vignettes supplied with this package.}
  \item{\code{new(...)}}{Create a new set of pivot table cells, specifying
  the field values documented above.}

  \item{\code{getCell(r, c))}}{Get the TableCell at the specified row and
  column coordinates in the pivot table.}
  \item{\code{setCell(r, c, cell))}}{Set the TableCell at the specified row
  and column coordinates in the pivot table.}
  \item{\code{getCells(specifyCellsAsList=FALSE, rowNumbers=NULL,
  columnNumbers=NULL, cellCoordinates=NULL)}}{Retrieve cells by a combination
  of row and/or column numbers.}
  \item{\code{findCells(variableNames=NULL, variableValues=NULL,
  totals="include", calculationNames=NULL, minValue=NULL, maxValue=NULL,
  exactValues=NULL, includeNull=TRUE, includeNA=TRUE)}}{Find cells matching
  the specified criteria.}
  \item{\code{getColumnWidths())}}{Retrieve the width of the longest value
  (in characters) in each column.}
  \item{\code{asMatrix(rawValue=TRUE))}}{Get a matrix containing all of the
  numerical values from the body of the pivot table (for rawValue=TRUE) or
  all of the formatted (i.e. character) values (for rawValue=FALSE).}
  \item{\code{asList())}}{Get a list representation of the pivot table
  cells.}
  \item{\code{asJSON()}}{Get a JSON representation of the pivot table cells.}
}
}

\examples{
# This class should only be created by the table.
# It is not intended to be created outside of the table.
}
\keyword{datasets}
